Implement DMLigatureCoordinator
