+++ DNM_iOS/DNM_iOS/ScoreViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import DNMModel
 
-// TODO: THIS IS BEING REFACTORED INTO FROM ENVIRONMENT (in-process: 2015-12-01)
+
 public class ScoreViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
     
     // MARK: - UI
@@ -57,6 +57,18 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
 
     // MARK: ScoreView Management
     
+    /**
+    Show the ScoreView with a PerformerID
+    
+    - parameter id: PerformerID
+    */
+    public func showScoreViewWithID(id: PerformerID) {
+        if let scoreView = scoreViewsByID[id] {
+            removeCurrentScoreView()
+            showScoreView(scoreView)
+        }
+    }
+    
     private func createScoreViews() {
         for viewerID in scoreViewIDs {
             let peerIDs = performerIDs.filter { $0 != viewerID }
@@ -64,13 +76,6 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
                 scoreModel: scoreModel, viewerID: viewerID, peerIDs: peerIDs
             )
             scoreViewsByID[viewerID] = scoreView
-        }
-    }
-    
-    public func showScoreViewWithID(id: PerformerID) {
-        if let scoreView = scoreViewsByID[id] {
-            removeCurrentScoreView()
-            showScoreView(scoreView)
         }
     }
     
@@ -103,8 +108,10 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
     
     private func manageColorMode() {
         view.backgroundColor = DNMColorManager.backgroundColor
-        
-        // wrap this up in a method
+        setViewSelectorTableViewBackground()
+    }
+    
+    private func setViewSelectorTableViewBackground() {
         let bgView = UIView()
         bgView.backgroundColor = DNMColorManager.backgroundColor
         viewSelectorTableView.backgroundView = bgView
@@ -112,18 +119,30 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
     
     // MARK: - PageLayer Navigation
     
+    /**
+    Go to the first page of the currently displayed ScoreView
+    */
     public func goToFirstPage() {
         currentScoreView?.goToFirstPage()
     }
-    
+
+    /**
+    Go to the last page of the currently displayed ScoreView
+    */
     public func goToLastPage() {
         currentScoreView?.goToLastPage()
     }
     
+    /**
+    Go to the next page of the currently displayed ScoreView
+    */
     public func goToNextPage() {
         currentScoreView?.goToNextPage()
     }
     
+    /**
+    Go to the previous page of the currently displayed ScoreView
+    */
     public func goToPreviousPage() {
         currentScoreView?.goToPreviousPage()
     }
@@ -170,45 +189,33 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
         return scoreViewIDs.count
     }
     
-    // TODO: CLEANUP, add Fields
+    // extend beyond title
     public func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath)
         -> UITableViewCell
     {
-        // PerformerID -- bold
-        // - InstrumentID (InstrumentType)
-        
         let cell = tableView.dequeueReusableCellWithIdentifier("scoreSelectorCell",
             forIndexPath: indexPath
         ) as! ScoreSelectorTableViewCell
         
         let viewerID = scoreViewIDs[indexPath.row]
+        
+        // do all this in ScoreSelectorTableViewCell implementation
         cell.identifier = viewerID
         cell.textLabel?.text = viewerID
-        
-        // SET COLOR IF VIEWER ID, or OMNI
-        
-        // manageStyling within ScoreSelectorTableViewCell
-        // color
+        setVisualAttributesOfTableViewCell(cell)
+        resizeViewSelectorTableView()
+        return cell
+    }
+
+    // do this in ScoreSelectorTableViewCell
+    private func setVisualAttributesOfTableViewCell(cell: UITableViewCell) {
         cell.textLabel?.textColor = UIColor.grayscaleColorWithDepthOfField(.Foreground)
         cell.backgroundColor = UIColor.grayscaleColorWithDepthOfField(DepthOfField.Background)
-        
 
-        
-        // make cleaner
         let selBGView = UIView()
         selBGView.backgroundColor = UIColor.grayscaleColorWithDepthOfField(.Middleground)
         cell.selectedBackgroundView = selBGView
-        
-        resizeViewSelectorTableView()
-        
-        return cell
     }
-    
-    /*
-    public func tableView(tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        return UIView(frame: CGRectZero)
-    }
-    */
     
     public override func prefersStatusBarHidden() -> Bool {
         return true
@@ -218,15 +225,4 @@ public class ScoreViewController: UIViewController, UITableViewDelegate, UITable
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
-    /*
-    // MARK: - Navigation
-    
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-    // Get the new view controller using segue.destinationViewController.
-    // Pass the selected object to the new view controller.
-    }
-    */
-    
 }
